refactored tests to show they are almost identical
